pyproject.toml
[project]
name = "3d-heritage-reconstruction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=5.49.1",
    "modal>=1.2.0",
    "pillow>=11.3.0",
]
